EFF-406 Move autoComplete prompt tests into Prompt.test
